requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
kaggle>=1.6.6
streamlit>=1.32.0
opencv-python>=4.9.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
tqdm>=4.66.0
Pillow>=10.2.0
numpy>=1.26.0
pandas>=2.2.0
ipykernel>=6.29.0
jupyter>=1.0.0
